Compare the NaN-check Double primitively in BoxesRunTime.equals
`x == x` on a boxed java.lang.Double makes the compiler synthesize a
call back into BoxesRunTime.equalsNumNum through the Any-comparator —
as a module-class-targeted static call that nothing emits a definition
for, failing the link. Unbox first so the comparison is primitive.

Written with LLM assistance (Claude Code).

+++ library-js-aux/src/scala/runtime/BoxesRunTime.scala
@@ -57,7 +57,7 @@ object BoxesRunTime {
     linkTimeIf(LinkingInfo.isWebAssembly) {
       if (x eq y) {
         x match {
-          case x: java.lang.Double => x == x // rejects NaN
+          case x: java.lang.Double => x.doubleValue() == x.doubleValue() // rejects NaN
           case _                   => true
         }
       } else equals2(x, y)
